refactor($theme-default): extract page components (#1427)
* refactor(default-theme): Extract PageEdit component and Extract PageNav component.
